Added Resizer component
